Refactoring: ContinuousSnapshotIntroductionPage, ContinuousSnapshotConfirmPage are separated to independent files.
